refact: Move apply to its own module
